issue-29 Create Clikt based factory for close programs
